safe use firebase config file

+++ src/components/cards/CoffeeCard.jsx
@@ -2,6 +2,7 @@ import { FaRegEye } from "react-icons/fa";
 import { MdEdit } from "react-icons/md";
 import { MdOutlineDelete } from "react-icons/md";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 
 
@@ -12,16 +13,33 @@ const CoffeeCard = ({coffee}) => {
 
     const handleDelete = (id) => {
         console.log(id);
-        fetch(`http://localhost:5000/coffee/${id}`,{
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data => {
-            console.log(data);
-            if(data. deletedCount > 0) {
-                alert('DELETED SUCCESSFULLY')
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/coffee/${id}`,{
+                    method: 'DELETE'
+                })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data);
+                    if(data. deletedCount > 0) {
+                        Swal.fire({
+                            title: "Deleted!",
+                            text: "Your file has been deleted.",
+                            icon: "success"
+                          });
+                    }
+                })
             }
-        })
+          });
+        
     }
     return (
         <div>
